+++ Plugins/SqlEnterpriseFormatter/formatwith.cpp
@@ -34,5 +34,16 @@ void FormatWithCommonTableExpression::formatInternal()
     if (cte->indexedColumns.size() > 0)
         withParDefLeft().withStatementList(cte->indexedColumns, "idxCols").withParDefRight();
 
-    withKeyword("AS").withParDefLeft().withStatement(cte->select).withParDefRight();
+    withKeyword("AS");
+    switch (cte->asMode) {
+        case SqliteWith::CommonTableExpression::ANY:
+            break;
+        case SqliteWith::CommonTableExpression::MATERIALIZED:
+            withKeyword("MATERIALIZED");
+            break;
+        case SqliteWith::CommonTableExpression::NOT_MATERIALIZED:
+            withKeyword("NOT").withKeyword("MATERIALIZED");
+            break;
+    }
+    withParDefLeft().withStatement(cte->select).withParDefRight();
 }